Move test code to dedicated test.
