+++ lib/3rd_party_adapters/command_line_interpreter.cpp
@@ -101,55 +101,3 @@ Command<ReturnType, Arguments...> makeCommand(
     };
     return command;
 }
-
-// ------------------- user code -----------------
-
-#include <iostream>
-
-static void testExtractWord(std::string &s)
-{
-    std::cout << "Next word\t = '" << (extractData<std::string>(s)).value() << "'" << std::endl;
-    std::cout << "Remainer\t = '" << s << "'" << std::endl;
-}
-
-static void foo()
-{
-    std::cout << "Hello World" << std::endl;
-}
-
-static void bar(const int n)
-{
-    std::cout << "Bar: " << n + 1 << std::endl;
-}
-
-static void two(const std::string s, const int n)
-{
-    std::cout << "Two: " << n << " " << s << std::endl;
-}
-
-int main()
-{
-
-    const auto myCommand = makeCommand("operate", std::make_tuple(), std::function(foo));
-    myCommand.parse("operate");
-
-    const Argument<int> barArg = {.labels = {"drinks"}, -42};
-    const auto myCommand2 = makeCommand("bar", std::make_tuple(&barArg), std::function(bar));
-    myCommand2.parse("bar drins 3");
-    myCommand2.parse("bar drinks 3");
-    myCommand2.parse("bar drinks ");
-    myCommand2.parse("bar drinks");
-    myCommand2.parse("bar");
-    myCommand2.parse("op drinks 3");
-    myCommand2.parse(" bar foo bar -55.5 drinks 3.9");
-    myCommand2.parse(" bar foo bar -55.5 drinks 4.9");
-
-    const Argument<std::string> object = {.labels = {"thing"}, "nothing"};
-    const Argument<int> number = {.labels = {"number"}, 0};
-    const auto myCommand3 = makeCommand("two", std::make_tuple(&object, &number), std::function(two));
-    myCommand3.parse("two");
-    myCommand3.parse("two thing=tree number 15");
-    myCommand3.parse("two number -1 thing car");
-
-    return 0;
-}
